Upgrading project to work with latest babel
It's considered best practice to not rely on global babel, etc, to build
a project. (citation needed). This specifies all dev dependencies
needed. This increases the npm install time but ultimately makes the
project more reliable.

Decorators are not fully supported in babel 6x so a plugin is required
to make them work.

It took a long time to track down all the bugs so please don't make me
do this again :)